--- a/libs/ts-sdk/src/sdk/cache/index.ts
+++ b/libs/ts-sdk/src/sdk/cache/index.ts
@@ -1,0 +1,17 @@
+export {
+  PoolDataCache,
+  CachedPoolMetadata,
+  CacheStats,
+  PoolCacheConfig,
+  DEFAULT_CACHE_CONFIG,
+  generatePoolCacheKey,
+} from "./pool-data-cache";
+
+export {
+  IPoolDataCache,
+  CacheError,
+  StaleDataError,
+  PoolNotFoundError,
+  CacheOptions,
+  DEFAULT_CACHE_OPTIONS,
+} from "./types";