--- a/libs/ts-sdk/src/sdk/cache/index.ts
+++ b/libs/ts-sdk/src/sdk/cache/index.ts
@@ -1,5 +1,6 @@
 export {
   PoolDataCache,
+  globalPoolDataCache,
   CachedPoolMetadata,
   CacheStats,
   PoolCacheConfig,