--- a/libs/ts-sdk/src/sdk/index.ts
+++ b/libs/ts-sdk/src/sdk/index.ts
@@ -2,4 +2,5 @@ export * from "./mira_amm";
 export * from "./readonly_mira_amm";
 export * from "./model";
 export * from "./errors";
+export * from "./cache";
 export {buildPoolId, getAssetId, getLPAssetId} from "./utils";