… libs/ts-sdk/src/sdk/__tests__/batch-preview.test.ts
… libs/ts-sdk/src/sdk/__tests__/caching-integration.test.ts
… libs/ts-sdk/src/sdk/__tests__/pool-preloading.test.ts
… libs/ts-sdk/src/sdk/__tests__/readonly-mira-amm-caching.test.ts
… libs/ts-sdk/src/sdk/cache/README.md
… libs/ts-sdk/src/sdk/cache/__tests__/cache-system.test.ts
… libs/ts-sdk/src/sdk/cache/__tests__/cached-pool-metadata.test.ts
… libs/ts-sdk/src/sdk/cache/__tests__/pool-data-cache.test.ts
… libs/web/src/hooks/__tests__/web-hooks-caching.test.tsx
… libs/web/src/hooks/get-swap-quotes-batch.ts
… libs/web/src/hooks/useSwapPreview.ts
… libs/web/src/hooks/useSwapPreviewV2.ts
… libs/web/src/hooks/useSwapRouter.ts