## Summary

Adds cacheing and preloading for pool reserves on the swap page

The overall effect of this is bringing down quote times from 0.5s->1s down to <100ms

This is a much snappier overall feeling

As a result of the faster fetching we can also reduce the debounce time for swap input to 100ms.

### Limitations

- [x] Still requires addition of unit tests

## Checklist

### 🧪 Tests & Quality

- [x] Added unit or integration tests for new functionality (using [Vitest](https://vitest.dev/))
- [x] All tests pass locally (`npx vitest`)

### ⚙️ Build & Tooling

- [x] Project starts successfully using `pnpm dev` or `pnpm start` (verifies build, dependencies,
      and runtime)
- [x] Ran `pnpm nx format:all` to ensure consistent code style
- [x] No obvious runtime or import errors (manual verification if needed)

### 🎨 UI (if applicable)

- [x] Storybook stories are added or updated (`pnpm storybook`)
- [x] Accessibility has been considered (e.g. labels, roles, focus management)
- [x] UI changes have been manually verified and screenshots are attached (if relevant)

### 🔍 Miscellaneous

- [x] No unrelated changes are included in this PR
- [x] Breaking changes are clearly documented, with migration steps if needed
- [x] Follows project conventions and folder structure

## Additional Note




<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **New Features**
  * Introduced advanced caching for pool data and swap quotes, improving performance and reducing network requests.
  * Added configurable cache options, including TTL, LRU eviction, and background preloading of pool data on route changes.
  * Enhanced swap preview responsiveness with faster input debouncing and caching.
  * Implemented pool metadata and fee caching with automatic expiration and error fallback.
  * Added route change detection to trigger selective pool data preloading.

* **Bug Fixes**
  * Improved error handling for swap quote fetching, including fallback logic when cached data is stale or unavailable.

* **Documentation**
  * Added comprehensive documentation for the new caching system and its configuration options.

* **Tests**
  * Added extensive test coverage for all caching features, including integration, preloading, cache statistics, and error scenarios.

* **Chores**
  * Minor code style and import optimizations for maintainability.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->